requirements/common.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==26.1.0
gunicorn==23.0.0
mysqlclient==2.2.4
packaging==24.1
pydot==3.0.1
PyJWT==2.8.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
